requirements.txt
apache-beam==2.48.0
apache-flink==1.19.0
apache-flink-libraries==1.19.0
avro-python3==1.10.2
certifi==2025.10.5
charset-normalizer==3.4.3
cloudpickle==2.2.1
confluent-kafka==2.12.0
crcmod==1.7
dill==0.3.1.1
dnspython==2.8.0
docopt==0.6.2
fastavro==1.12.1
fasteners==0.20
find_libpython==0.5.0
grpcio==1.75.1
hdfs==2.7.3
httplib2==0.22.0
idna==3.11
numpy==1.24.4
objsize==0.6.1
orjson==3.11.3
pandas==2.3.3
pemja==0.4.1
proto-plus==1.26.1
protobuf==4.23.4
psutil==7.1.0
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.9.18
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
six==1.17.0
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zstandard==0.25.0
